refactor(ci): extract the chainguard-fips and master/support job groups
Both were private methods of the pull-request workflow and therefore unusable anywhere else.
They move verbatim to workflows/groups/, next to the e2e and backend image groups.

masterAndSupportJobs stays heterogeneous on purpose: it still mixes the chainguard images, the
community build check, the SaaS trigger, the Helm and Maven publications and the two
deployments. Splitting it along those lines reorders the job declarations, which is a change of
behaviour rather than a move, so it is left for a dedicated change.

workflow-pull-requests.ts drops from 961 to 687 lines. Generated configuration is byte-for-byte
identical: all 50 fixtures are untouched.
