+++ .circleci/ci/src/workflows/groups/chainguard-fips-jobs.ts
@@ -1,0 +1,80 @@
+/*
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { Config, workflow } from '../../circleci-config';
+import { BuildDockerChainguardFipsImageJob } from '../../jobs';
+import { CircleCIEnvironment } from '../../pipelines';
+import { config } from '../../config';
+
+/**
+ * The chainguard-FIPS component images, published to the private Azure registry only.
+ */
+export function chainguardFipsJobs(dynamicConfig: Config, environment: CircleCIEnvironment): workflow.WorkflowJob[] {
+  const buildDockerChainguardFipsImageJob = BuildDockerChainguardFipsImageJob.create(dynamicConfig, environment, false);
+  dynamicConfig.addJob(buildDockerChainguardFipsImageJob);
+
+  return [
+    new workflow.WorkflowJob(buildDockerChainguardFipsImageJob, {
+      context: config.jobContext,
+      name: `Build APIM Management API chainguard-fips docker image`,
+      requires: ['Build backend'],
+      'apim-project': config.components.managementApi.project,
+      'apim-project-workdir': config.components.managementApi.distribution,
+      'docker-context': 'target',
+      'docker-image-name': config.components.managementApi.image,
+      'docker-fips-base-image': config.docker.fipsJavaBaseImage,
+    }),
+    new workflow.WorkflowJob(buildDockerChainguardFipsImageJob, {
+      context: config.jobContext,
+      name: `Build APIM Gateway chainguard-fips docker image`,
+      requires: ['Build backend'],
+      'apim-project': config.components.gateway.project,
+      'apim-project-workdir': config.components.gateway.distribution,
+      'docker-context': 'target',
+      'docker-image-name': config.components.gateway.image,
+      'docker-fips-base-image': config.docker.fipsJavaBaseImage,
+    }),
+    new workflow.WorkflowJob(buildDockerChainguardFipsImageJob, {
+      context: config.jobContext,
+      name: `Build APIM Console chainguard-fips docker image`,
+      requires: ['Build APIM Console'],
+      'apim-project': config.components.console.project,
+      'apim-project-workdir': config.components.console.workdir,
+      'docker-context': '.',
+      'docker-image-name': config.components.console.image,
+      'docker-fips-base-image': config.docker.fipsNginxBaseImage,
+    }),
+    new workflow.WorkflowJob(buildDockerChainguardFipsImageJob, {
+      context: config.jobContext,
+      name: `Build APIM Portal chainguard-fips docker image`,
+      requires: ['Build APIM Portal'],
+      'apim-project': config.components.portal.project,
+      'apim-project-workdir': config.components.portal.workdir,
+      'docker-context': '.',
+      'docker-image-name': config.components.portal.image,
+      'docker-fips-base-image': config.docker.fipsNginxBaseImage,
+    }),
+    new workflow.WorkflowJob(buildDockerChainguardFipsImageJob, {
+      context: config.jobContext,
+      name: `Build Gamma Console chainguard-fips docker image`,
+      requires: ['Build Gamma Console'],
+      'apim-project': config.components.gamma.project,
+      'apim-project-workdir': config.components.gamma.workdir,
+      'docker-context': '.',
+      'docker-image-name': config.components.gamma.image,
+      'docker-fips-base-image': config.docker.fipsNginxBaseImage,
+    }),
+  ];
+}
